+++ public/js/administratepaper/view.js
@@ -7,10 +7,7 @@ function getTags() {
         if (paper.title[locale]) {
             paper_title = paper.title[locale];
         } else {
-            for (let i in paper.title) {
-                paper_title = paper.title[i];
-                break;
-            }
+            paper_title = paper.title[0];
         }
     } else {
         paper_title = paper.title;