+++ application/modules/journal/controllers/AdministratepaperController.php
@@ -120,6 +120,8 @@ class AdministratepaperController extends PaperDefaultController
             // La liste des articles, après filtrage
             $papers = $review->getPapers($settings, false, true);
 
+            $isCoiEnabled = $review->getSetting(Episciences_Review::SETTING_SYSTEM_IS_COI_ENABLED);
+
             foreach ($papers as &$paper) {
                 $paper->loadSubmitter(false);
                 $paper->getEditors();
@@ -134,6 +136,10 @@ class AdministratepaperController extends PaperDefaultController
                 ); // environ 1s
 
                 $paper->setRevisionDeadline();
+
+                if($isCoiEnabled){
+                    $paper->getConflicts(true);
+                }
 
             }
 
@@ -289,6 +295,9 @@ class AdministratepaperController extends PaperDefaultController
             // liste des articles à afficher
             $papers = $user->loadAssignedPapers($settings, true);
 
+            $isCoiEnabled = $review->getSetting(Episciences_Review::SETTING_SYSTEM_IS_COI_ENABLED);
+
+
             /** @var Episciences_Paper $paper */
             foreach ($papers as &$paper) {
                 $paper->loadSubmitter(false);
@@ -297,6 +306,11 @@ class AdministratepaperController extends PaperDefaultController
                 $paper->getReviewers([Episciences_User_Assignment::STATUS_ACTIVE, Episciences_User_Assignment::STATUS_PENDING], true);
                 $paper->setRevisionDeadline();
             }
+
+            if($isCoiEnabled){
+                $paper->getConflicts(true);
+            }
+
             unset($paper);
             $tbody = (count($papers) > 0) ?
                 $this->view->partial('administratepaper/datatable_list.phtml', [