+++ public/js/functions.js
@@ -677,7 +677,7 @@ function updateDeadlineTag(body, tagName, date, locale = 'en') {
     let content = body.getContent();
     let replace = '<span class="revision_deadline">' + translate('dès que possible', locale) + '</span>';
 
-    if (isISOdate(date) && isValidDate(date)) {
+    if (isISOdate(date) && isValidDate(date) && !isBackDated(date, locale)) {
 
         let localeDate = getLocaleDate(date, {
             language: locale,
@@ -803,6 +803,6 @@ function enableModalSubmitButton(){
     }
 }
 
-
-
-
+function isBackDated(input, locale){
+    return new Date().toLocaleDateString(locale) > new Date(input).toLocaleDateString(locale);
+}